Create conclusion.js
Code made to check if there are any discounts by checking the number of guests along with the dishes name and price.
Then check at the by the required conditions of a boolean and the specific number argument. So different inputs result in different results by the double checking condition of the getDiscount function.